pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aars"
version = "0.1.0"
description = "Agentic Adaptive Retrieval System — dynamic strategy selection for RAG"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    {name = "lekhanpro"},
]
keywords = ["rag", "retrieval-augmented-generation", "llm", "agents", "adaptive-retrieval"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "anthropic>=0.40.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    "rank-bm25>=0.2.2",
    "networkx>=3.3",
    "spacy>=3.7.0",
    "pydantic-settings>=2.5.0",
    "python-multipart>=0.0.12",
    "httpx>=0.27.0",
    "structlog>=24.4.0",
    "PyMuPDF>=1.24.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "httpx>=0.27.0",
]
bench = [
    "datasets>=3.0.0",
    "pandas>=2.2.0",
    "matplotlib>=3.9.0",
    "scipy>=1.14.0",
    "tabulate>=0.9.0",
]
ui = [
    "streamlit>=1.38.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
